+++ PayFacMpSDK/PayFacMpSDK/legalEntityUpdateRequest.cs
@@ -63,7 +63,7 @@ namespace PayFacMpSDK
             }
             if(legalEntityOwnershipTypeSpecified) xmlBuilder.Append("<legalEntityOwnershipType>" + legalEntityOwnershipType + "</legalEntityOwnershipType>");
             if(yearsInBusiness != null) xmlBuilder.Append("<yearsInBusiness>" + yearsInBusiness + "</yearsInBusiness>");
-            
+            if (pciLevelIntValue != 0) xmlBuilder.Append("<pciLevel>" + pciLevelIntValue + "</pciLevel>");
             xmlBuilder.Append("</legalEntityUpdateRequest>");
             Console.WriteLine(xmlBuilder.ToString());
             return xmlBuilder.ToString();