+++ PayFacMpSDK/PayFacMpSDK/subMerchantUpdateRequest.cs
@@ -91,6 +91,21 @@ namespace PayFacMpSDK
                 methodOfPayments.Serialize(xmlBuilder);
                 xmlBuilder.Append("</methodOfPayments>");
             }
+            if (countryOfOrigin != null)
+            {
+                xmlBuilder.Append("<countryOfOrigin>" + countryOfOrigin + "</countryOfOrigin>");
+            }
+            if (revenueBoost != null)
+            {
+                xmlBuilder.Append("<revenueBoost enabled =\"" + revenueBoost.enabled.ToString().ToLower() + "\">");
+                xmlBuilder.Append("</revenueBoost>");
+            }
+            if (complianceProducts != null)
+            {
+                xmlBuilder.Append("<complianceProducts>");
+                complianceProducts.Serialize(xmlBuilder);
+                xmlBuilder.Append("</complianceProducts>");
+            }
             xmlBuilder.Append("</subMerchantUpdateRequest>");
             Console.WriteLine(xmlBuilder.ToString());
             return xmlBuilder.ToString();